requirements.txt

line-bot-sdk
Flask
Flask-SocketIO
Flask-Login
pymongo
uwsgi
